requirements.txt
# AI Writing Agent - Enterprise Open Source Edition
# Requirements

# Core
langchain>=0.2.0
langchain-core>=0.2.0

# Providers
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
langchain-ollama>=0.1.0

# API Server
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.0.0
python-multipart>=0.0.6

# HTTP Client
httpx>=0.26.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Document Processing
pypdf>=4.0.0
python-docx>=1.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Development
ruff>=0.2.0
black>=24.0.0
pre-commit>=3.6.0

# Optional: Local Embeddings
# sentence-transformers>=2.2.0
